requirements.txt
Django>=4.2
gunicorn
pymodbus
paho-mqtt
psycopg2-binary  
whitenoise
requests
openleadr

